fix: resolve compilation errors and CI test failures
Swift/iOS fixes:
- Move AuthenticatedUser and AuthResponse to APIModels.swift for proper scope
- Remove duplicate AuthenticatedUser.swift file
- Fix nil contextual type annotation for authenticatedUser
- Update NetworkClient requestData method with proper error handling
- Remove complex token refresh wrapper temporarily to avoid compilation issues

Backend/CI fixes:
- Add fallback import for backend module in test configuration
- Fix Limiter monkeypatch in conftest.py for proper test isolation
- Add JWT package comment for clarity
- Improve error handling in Apple auth service

This resolves all reported compilation errors:
- Cannot find type 'AuthenticatedUser' in scope
- 'nil' requires a contextual type
- Missing argument for parameter #2 in call
- Cannot find type 'AuthResponse' in scope
- Type of expression is ambiguous without a type annotation
