+++ backend/tests/conftest.py
@@ -9,16 +9,25 @@ for path in (ROOT_DIR, BACKEND_DIR):
     if path not in sys.path:
         sys.path.insert(0, path)
 
-from backend import app as app_module
+try:
+    from backend import app as app_module
+except ImportError:
+    # Fallback for CI environment
+    import app as app_module
 
-class TestLimiter(app_module.Limiter):
+# Import Limiter directly for testing
+from flask_limiter import Limiter
+
+class TestLimiter(Limiter):
     def __init__(self, *args, **kwargs):
         kwargs['default_limits'] = ["2 per minute"]
         super().__init__(*args, **kwargs)
 
 @pytest.fixture
 def app(monkeypatch):
-    monkeypatch.setattr(app_module, 'Limiter', TestLimiter)
+    # Patch the Limiter import in the app module
+    import flask_limiter
+    monkeypatch.setattr(flask_limiter, 'Limiter', TestLimiter)
     application = app_module.create_app()
     application.config['TESTING'] = True
     return application
